+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/util/AutoLinkifyTest.kt
@@ -2,12 +2,6 @@ package de.connect2x.trixnity.messenger.util
 
 import de.connect2x.trixnity.messenger.util.html.AutoLinkifyVisitor
 import de.connect2x.trixnity.messenger.util.html.HtmlNode
-import de.connect2x.trixnity.messenger.util.html.MatrixMentionVisitor
-import io.ktor.http.*
-import net.folivo.trixnity.core.model.Mention
-import net.folivo.trixnity.core.model.RoomAliasId
-import net.folivo.trixnity.core.model.UserId
-import kotlin.math.exp
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -24,35 +18,48 @@ class AutoLinkifyTest {
     @Test
     fun linkifiesLinksInText() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "https://example.com/",
-                    ), listOf(
-                        HtmlNode.TextContent("https://example.com/"),
-                    )),
-                    HtmlNode.TextContent(" Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "https://example.com/",
+                                ), listOf(
+                                    HtmlNode.TextContent("https://example.com/"),
+                                )
+                            ),
+                            HtmlNode.TextContent(" Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. https://example.com/ Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."),
         )
     }
+
     // Mentions
     @Test
     fun shouldReplaceBasicUserMention() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Hallo "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "matrix:u/user:acme.com",
-                    ), listOf(
-                        HtmlNode.TextContent("@user:acme.com"),
-                    )),
-                    HtmlNode.TextContent("! How are you?"),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Hallo "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "matrix:u/user:acme.com",
+                                ), listOf(
+                                    HtmlNode.TextContent("@user:acme.com"),
+                                )
+                            ),
+                            HtmlNode.TextContent("! How are you?"),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Hallo @user:acme.com! How are you?"),
         )
     }
@@ -60,17 +67,23 @@ class AutoLinkifyTest {
     @Test
     fun shouldReplaceBasicRoomAliasMention() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Hallo "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "matrix:r/awesome-room:acme.com",
-                    ), listOf(
-                        HtmlNode.TextContent("#awesome-room:acme.com"),
-                    )),
-                    HtmlNode.TextContent("! How are you?"),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Hallo "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "matrix:r/awesome-room:acme.com",
+                                ), listOf(
+                                    HtmlNode.TextContent("#awesome-room:acme.com"),
+                                )
+                            ),
+                            HtmlNode.TextContent("! How are you?"),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Hallo #awesome-room:acme.com! How are you?"),
         )
     }
@@ -78,17 +91,23 @@ class AutoLinkifyTest {
     @Test
     fun shouldReplaceUriUserMention() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Hallo "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "matrix:u/user:acme.com?action=chat",
-                    ), listOf(
-                        HtmlNode.TextContent("@user:acme.com"),
-                    )),
-                    HtmlNode.TextContent("! How are you?"),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Hallo "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "matrix:u/user:acme.com?action=chat",
+                                ), listOf(
+                                    HtmlNode.TextContent("@user:acme.com"),
+                                )
+                            ),
+                            HtmlNode.TextContent("! How are you?"),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Hallo matrix:u/user:acme.com?action=chat! How are you?"),
         )
     }
@@ -96,17 +115,23 @@ class AutoLinkifyTest {
     @Test
     fun shouldReplaceLinkUserMention() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Hallo "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "https://matrix.to/#/%40alice%3Aexample.org",
-                    ), listOf(
-                        HtmlNode.TextContent("@alice:example.org"),
-                    )),
-                    HtmlNode.TextContent("! How are you?"),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Hallo "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "https://matrix.to/#/%40alice%3Aexample.org",
+                                ), listOf(
+                                    HtmlNode.TextContent("@alice:example.org"),
+                                )
+                            ),
+                            HtmlNode.TextContent("! How are you?"),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Hallo https://matrix.to/#/%40alice%3Aexample.org! How are you?"),
         )
     }
@@ -114,23 +139,31 @@ class AutoLinkifyTest {
     @Test
     fun shouldReplaceMultipleMentions() {
         assertEquals(
-            expected = HtmlNode.HtmlElement("#root", emptyMap(), listOf(
-                HtmlNode.HtmlElement("span", emptyMap(), listOf(
-                    HtmlNode.TextContent("Hallo "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "matrix:u/user:acme.com",
-                    ), listOf(
-                        HtmlNode.TextContent("@user:acme.com"),
-                    )),
-                    HtmlNode.TextContent("! How are you? Want to meet up with "),
-                    HtmlNode.HtmlElement("a", mapOf(
-                        "href" to "matrix:u/user2:acme.com",
-                    ), listOf(
-                        HtmlNode.TextContent("@user2:acme.com"),
-                    )),
-                    HtmlNode.TextContent("?"),
-                )),
-            )),
+            expected = HtmlNode.HtmlElement(
+                "#root", emptyMap(), listOf(
+                    HtmlNode.HtmlElement(
+                        "span", emptyMap(), listOf(
+                            HtmlNode.TextContent("Hallo "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "matrix:u/user:acme.com",
+                                ), listOf(
+                                    HtmlNode.TextContent("@user:acme.com"),
+                                )
+                            ),
+                            HtmlNode.TextContent("! How are you? Want to meet up with "),
+                            HtmlNode.HtmlElement(
+                                "a", mapOf(
+                                    "href" to "matrix:u/user2:acme.com",
+                                ), listOf(
+                                    HtmlNode.TextContent("@user2:acme.com"),
+                                )
+                            ),
+                            HtmlNode.TextContent("?"),
+                        )
+                    ),
+                )
+            ),
             actual = parseLinks("Hallo @user:acme.com! How are you? Want to meet up with @user2:acme.com?"),
         )
     }
@@ -161,13 +194,16 @@ class AutoLinkifyTest {
             )
         }
     }
+
     @Test
     fun shouldFormatRegularLink() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
                                     "href" to "https://matrix.org",
@@ -187,9 +223,11 @@ class AutoLinkifyTest {
     fun shouldFormatUrlWithEscapedAmpersand() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
                                     "href" to "https://duckduckgo.com/?q=html+escaping+ampersand&amp;ia=web",
@@ -209,9 +247,11 @@ class AutoLinkifyTest {
     fun shouldFormatUrlWithSemicolon() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
                                     "href" to "https://exampleformytest.com/bla;blubb",
@@ -231,9 +271,11 @@ class AutoLinkifyTest {
     fun shouldFormatUrlWithComma() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
                                     "href" to "https://osmand.net/map/?pin=50.89774,13.69089#19/51.05483/13.74711",
@@ -253,9 +295,11 @@ class AutoLinkifyTest {
     fun shouldIgnoreExclamationMarkAtEndOfLink() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("I think you could really like "),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -277,9 +321,11 @@ class AutoLinkifyTest {
     fun shouldIgnoreParenthesis() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("The website Graphemica ("),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -301,9 +347,11 @@ class AutoLinkifyTest {
     fun shouldIgnoreColonAtTheEndOfLink() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("The best thing about "),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -325,9 +373,11 @@ class AutoLinkifyTest {
     fun shouldIgnoreQuestionMarkAtTheEndOfLink() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("Do you know "),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -349,9 +399,11 @@ class AutoLinkifyTest {
     fun shouldIgnorePeriodAtTheEndOfLink() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("I thought about "),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -373,9 +425,11 @@ class AutoLinkifyTest {
     fun shouldAllowUnicodeSymbols() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
                                     "href" to "https://graphemica.com/»",
@@ -395,9 +449,11 @@ class AutoLinkifyTest {
     fun shouldAllowParenthesisInUrls() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("Have you heard about "),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(
@@ -419,9 +475,11 @@ class AutoLinkifyTest {
     fun shouldRemovePunctuationAroundUrlsWithParenthesises() {
         assertEquals(
             expected = HtmlNode.HtmlElement(
-                "#root", emptyMap(), listOf(
+                "#root", emptyMap(),
+                listOf(
                     HtmlNode.HtmlElement(
-                        "span", emptyMap(), listOf(
+                        "span", emptyMap(),
+                        listOf(
                             HtmlNode.TextContent("Have you heard about Matrix ("),
                             HtmlNode.HtmlElement(
                                 "a", mapOf(