+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/ReportMessageViewModelTest.kt
@@ -32,16 +32,18 @@ import net.folivo.trixnity.core.model.EventId
 import net.folivo.trixnity.core.model.RoomId
 import net.folivo.trixnity.core.model.UserId
 import net.folivo.trixnity.core.model.events.ClientEvent
+import net.folivo.trixnity.core.model.events.RoomEventContent
 import net.folivo.trixnity.core.model.events.m.room.MemberEventContent
 import net.folivo.trixnity.core.model.events.m.room.Membership
 import net.folivo.trixnity.core.model.events.m.room.RoomMessageEventContent
 import org.koin.dsl.koinApplication
 import org.koin.dsl.module
+import kotlin.reflect.KClass
 import kotlin.test.Test
 import kotlin.time.Duration.Companion.milliseconds
 
 class ReportMessageViewModelTest {
-    private val roomId = RoomId("room1", "localhost")
+    private val roomId = RoomId("!room1")
 
     private val ourUserId = UserId("bob", "localhost")
 
@@ -92,7 +94,7 @@ class ReportMessageViewModelTest {
         every { matrixClientServerApiClientMock.room } returns roomsApiClientMock
 
         canSendEventMocker = every {
-            userServiceMock.canSendEvent(any(), any())
+            userServiceMock.canSendEvent(any(), any<KClass<out RoomEventContent>>())
         }
 
         canSendEventMocker returns flowOf(true)