+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New rich text renderer
 - Show correct feedback when getting no search results in the room list
 - Labels for text fields are always on
-
+- Adapt to trixnity References-API changes
 
 ### Deprecated